fix/devide file modal function
