+++ frontend/src/Chat/components/ChatList.tsx
@@ -2,9 +2,10 @@ import * as React from 'react'
 import '../assets/styles.css'
 import { useLocation, useNavigate } from 'react-router-dom'
 import { useEffect, type ReactElement } from 'react'
-import { Alert, Button, Dropdown, DropdownButton, Modal } from 'react-bootstrap'
+import { Alert, Button } from 'react-bootstrap'
 import 'bootstrap/dist/css/bootstrap.min.css'
 import axios from 'axios'
+import { ChatListModal } from './ChatListModal'
 
 axios.defaults.baseURL = process.env.REACT_APP_BACKEND_HTTP_BASE_URL
 
@@ -15,102 +16,6 @@ const _publicIconURL: string =
   'https://iconbox.fun/wp/wp-content/uploads/lock_open_24.png'
 const privateIconURL: string =
   'https://iconbox.fun/wp/wp-content/uploads/lock_24.png'
-
-const PublicSelectDropdownButton = (props: {
-  setIsPublic: React.Dispatch<React.SetStateAction<boolean>>
-}): ReactElement => {
-  const [title, setTitle] = React.useState<string>('public')
-
-  return (
-    <DropdownButton
-      id="dropdown-basic-button"
-      variant="info"
-      title={title}
-      onSelect={(
-        op: string | null,
-        _e: React.SyntheticEvent<unknown>
-      ): void => {
-        setTitle(op ?? 'select')
-        switch (op) {
-          case 'public':
-            props.setIsPublic(true)
-            break
-          case 'private':
-            props.setIsPublic(false)
-            break
-        }
-      }}
-    >
-      <Dropdown.Item eventKey="public">public</Dropdown.Item>
-      <Dropdown.Item eventKey="private">private</Dropdown.Item>
-    </DropdownButton>
-  )
-}
-
-const CreateRoomModal = (props: {
-  user: User
-  showCreateRoomModal: boolean
-  handleModalClose: () => void
-}): ReactElement => {
-  const [newRoomName, setNewRoomName] = React.useState<string>('')
-  const [isPublic, setIsPublic] = React.useState<boolean>(true)
-
-  const handleCreateRoom = (): void => {
-    const requestData: ChatRoom = {
-      name: newRoomName,
-      created_by: props.user.id,
-      isPublic
-    }
-    axios
-      .post<ChatRoom>('/chatRoom', requestData)
-      .then((_response) => {
-        props.handleModalClose()
-      })
-      .catch((reason) => {
-        alert('エラーです！')
-        console.log(reason)
-      })
-  }
-
-  return (
-    <>
-      <Modal show={props.showCreateRoomModal} onHide={props.handleModalClose}>
-        <Modal.Header closeButton>
-          <Modal.Title>create new Room</Modal.Title>
-        </Modal.Header>
-
-        <Modal.Body>
-          <>
-            new room:
-            <input
-              type="text"
-              value={newRoomName}
-              onChange={(e) => {
-                setNewRoomName(e.target.value)
-              }}
-            />
-            <PublicSelectDropdownButton
-              setIsPublic={setIsPublic}
-            ></PublicSelectDropdownButton>
-          </>
-        </Modal.Body>
-        <Modal.Footer>
-          <Button variant="secondary" onClick={props.handleModalClose}>
-            Close
-          </Button>
-          <Button
-            variant="primary"
-            onClick={() => {
-              handleCreateRoom()
-            }}
-          >
-            Create
-          </Button>
-        </Modal.Footer>
-      </Modal>
-    </>
-  )
-}
 
 export function ChatList(): ReactElement {
   const [user, setUser] = React.useState<User>({ id: 0, name: '' })
@@ -180,11 +85,11 @@ export function ChatList(): ReactElement {
   return (
     <>
       {alertElement}
-      <CreateRoomModal
+      <ChatListModal
         user={user}
         showCreateRoomModal={showCreateRoomModal}
         handleModalClose={handleModalClose}
-      ></CreateRoomModal>
+      ></ChatListModal>
       <div className="Chat">
         <p>Enter your name and Move on to a chat room.</p>
         <p>
